:truck: Move commands from mod.ts to ./commands/*.ts
